moved summary components
